// src/modules/com/cliente/contato/dto/update-contato.dto.ts
export class UpdateContatoDto {
  bin_codcli?: number;
  vch_nomctocli?: string;
  vch_cgccpfcli?: string;
  vch_idtctocli?: string;
  chr_estctocli?: string;
  vch_cidctocli?: string;
  vch_baictocli?: string;
  vch_cepctocli?: string;
  vch_endctocli?: string;
  int_numendctocli?: number;
  vch_cplendctocli?: string;
  vch_numtelctocli?: string;
  vch_numcelctocli?: string;
  vch_crectocli?: string;
  dat_datnscctocli?: string;
  vch_obsctocli?: string;
}